+++ Mailer/extensions/mailer-module/src/migrations/20230217084810_unique_configs_name.php
@@ -1,0 +1,14 @@
+<?php
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class UniqueConfigsName extends AbstractMigration
+{
+    public function change(): void
+    {
+        $this->table('configs')
+            ->addIndex('name', ['unique' => true])
+            ->update();
+    }
+}