+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added ability to track RTM parameters in the `/api/v1/users/subscribe` API. remp/remp#1237
 - **IMPORTANT**: Changed primary key from `int` to `bigint` for `mail_user_subscriptions` table. remp/remp#1187
   - This migration is a two-step process that requires your manual action - running `mail:migrate-user-subscriptions-and-variants` in the off-peak hours. Since some tables are very exposed and cannot be locked for more than a couple of seconds, we decided to migrate the data into the new table manually and keep the old and new table in sync. Based on the amount of your data, the migration can take hours.
+- Added support for standalone HTTP webhook signing key. remp/remp#1232
+  - Mailgun used to use domains API key to sign the requests, however it currently is a separate signing key.
 
 ## [1.1.0] - 2023-01-27